Implement project sharing

+++ src/server/modules/project/controller/controller.js
@@ -70,6 +70,18 @@ async function updateProject(req, res) {
         return res.status(400).json(errorResponse);
     }
 
+    const [hasPermissionToUpdate, permissionErr] = await ProjectService.hasPermission(id, req.user.id, 'PROJECT', 'UPDATE');
+
+    if (permissionErr) {
+        permissionErr.forEach(e => errorResponse.addError(e.message, ''));
+        return res.status(400).json(errorResponse);
+    }
+
+    if (!hasPermissionToUpdate) {
+        errorResponse.addError('Forbidden.', '');
+        return res.status(403).send(errorResponse);
+    }
+
     const [project, err] = await ProjectService.updateProject(id, { name });
 
     if (err) {
@@ -84,8 +96,21 @@ async function updateProject(req, res) {
 
 async function deleteProject(req, res) {
     const successResponse = new Response.success();
+    const errorResponse = new Response.error();
 
     const id = req.params.id;
+
+    const [hasPermissionToDelete, err] = await ProjectService.hasPermission(id, req.user.id, 'PROJECT', 'DELETE');
+
+    if (err) {
+        err.forEach(e => errorResponse.addError(e.message, ''));
+        return res.status(400).json(errorResponse);
+    }
+
+    if (!hasPermissionToDelete) {
+        errorResponse.addError('Forbidden.', '');
+        return res.status(403).send(errorResponse);
+    }
 
     const project = await ProjectService.deleteProject(id);
 
@@ -94,7 +119,87 @@ async function deleteProject(req, res) {
     res.json(successResponse);
 }
 
+async function shareProject(req, res) {
+    const errorResponse = new Response.error();
+
+    const requestingUserID = req.user.id;
+    const id = req.params.id;
+    const { 
+        user_id: userID, 
+        can_read: hasReadAccess, 
+        can_write: hasWriteAccess
+    } = req.body;
+
+    if (requestingUserID === userID) {
+        errorResponse.addError('User cannot modify its own access to a project.', '');
+        return res.status(403).json(errorResponse);
+    }
+
+    const [hasPermissionToShare, err] = await ProjectService.hasPermission(id, requestingUserID, 'PROJECT', 'SHARE');
+
+    if (err) {
+        err.forEach(e => errorResponse.addError(e.message, ''));
+        return res.status(400).json(errorResponse);
+    }
+
+    if (!hasPermissionToShare) {
+        errorResponse.addError('The user does not have the permission to share the project.', '');
+        return res.status(403).json(errorResponse);
+    }
+
+    await ProjectService.shareProject(id, userID, hasReadAccess, hasWriteAccess);
+
+    res.sendStatus(200);
+}
+
+async function getProjectUsers(req, res) {
+    const successResponse = new Response.success();
+
+    const id = req.params.id;
+
+    const [project] = await ProjectService.getProjectUsers(id);
+
+    successResponse.data = ProjectViewModels.project(project);
+
+    res.json(successResponse);
+}
+
+async function revokeProjectUserAccess(req, res) {
+    const errorResponse = new Response.error();
+    const { id: projectID, userID } = req.params;
+    const requestingUserID = req.user.id;
+    
+    if (requestingUserID === userID) {
+        errorResponse.addError('User cannot modify its own access to a project.', '');
+        return res.status(403).json(errorResponse);
+    }
+
+    const [hasPermissionToShare, permissionErr] = await ProjectService.hasPermission(projectID, requestingUserID, 'PROJECT', 'SHARE');
+
+    if (permissionErr) {
+        permissionErr.forEach(e => errorResponse.addError(e.message, ''));
+        return res.status(400).json(errorResponse);
+    }
+
+    if (!hasPermissionToShare) {
+        errorResponse.addError('The user does not have the permission to revoke access of a user.', '');
+        return res.status(403).json(errorResponse);
+    }
+
+    const [, err] = await ProjectService.revokeProjectUserAccess(projectID, userID);
+
+    if (err) {
+        err.forEach(e => errorResponse.addError(e.message, ''));
+        return res.status(400).json(errorResponse);
+    }
+
+    res.sendStatus(200);
+}
+
 exports.getProject = getProject;
 exports.createProject = createProject;
 exports.updateProject = updateProject;
 exports.deleteProject = deleteProject;
+exports.shareProject = shareProject;
+exports.getProjectUsers = getProjectUsers;
+exports.revokeProjectUserAccess = revokeProjectUserAccess;
